Blocks folder refactoring Plugins load improve
