+++ main.js
@@ -27,7 +27,7 @@ let program = require('commander');
 
 program
     .version(version)
-    .description(' iZ3 blockchain core.')
+    .description(' iZ3 - IZZZIO blockchain core.')
     .option('-a, --autofix', 'Fix saved chain if possible. WARNING: You can lose important data')
     .option('--clear', 'Clear all saved chain and deletes wallet. WARNING: You can lose important data')
     .option('--clear-db', 'Clear all saved chain and calculated wallets.')
