+++ modules/blocksModels/candyData.js
