+++ modules/blocksModels/walletRegister.js
